VCV-96: annotation panel shows morph ID annotation form for every specimen
