+++ next.config.ts
@@ -1,7 +1,7 @@
 import type { NextConfig } from 'next';
 
 const nextConfig: NextConfig = {
-  /* config options here */
+  transpilePackages: ['@radix-ui/react-hover-card'],
   images: {
     remotePatterns: [
       {
