+++ src/features/annotation/api/index.ts
@@ -2,5 +2,4 @@ export * from './annotation-keys';
 export * from './get-annotation-tasks';
 export * from './get-annotation-task-progress';
 export * from './get-annotations';
-export * from './get-specimen';
 export * from './update-annotation';
