+++ src/features/annotation/components/task-detail/constants.ts
@@ -1,0 +1,2 @@
+export const ANNOTATION_TASK_DETAIL_LAYOUT_CLASS =
+  'mx-auto grid h-full w-full max-w-[1800px] gap-6 p-6 md:grid-cols-[minmax(0,1.4fr)_minmax(380px,1fr)_minmax(380px,1fr)]';
