+++ src/features/annotation/components/task-detail/loading-skeleton.tsx
@@ -1,12 +1,7 @@
 import { Card, CardContent, CardFooter, CardHeader } from '@/ui/card';
 import { Skeleton } from '@/ui/skeleton';
+import { ANNOTATION_TASK_DETAIL_LAYOUT_CLASS } from './constants';
 
-const getLayoutClass = (shouldProcessFurther: boolean) => {
-  if (shouldProcessFurther) {
-    return 'mx-auto grid h-full w-full max-w-[1800px] gap-6 p-6 md:grid-cols-[minmax(0,1.4fr)_minmax(380px,1fr)_minmax(380px,1fr)]';
-  }
-  return 'mx-auto grid h-full w-full max-w-6xl gap-6 p-6 md:grid-cols-[1fr_1fr]';
-};
 const statusGridClass =
   'grid gap-2 justify-items-center sm:grid-cols-2 sm:justify-items-stretch lg:grid-cols-3';
 const statusCardClass =
@@ -22,17 +17,9 @@ const StatusSkeleton = () => (
   </div>
 );
 
-interface AnnotationTaskDetailSkeletonProps {
-  shouldProcessFurther?: boolean;
-}
-
-export function AnnotationTaskDetailSkeleton({
-  shouldProcessFurther = false,
-}: AnnotationTaskDetailSkeletonProps) {
-  const layoutClass = getLayoutClass(shouldProcessFurther);
-
+export function AnnotationTaskDetailSkeleton() {
   return (
-    <div className={layoutClass}>
+    <div className={ANNOTATION_TASK_DETAIL_LAYOUT_CLASS}>
       <Card className="flex h-full flex-col overflow-hidden shadow-sm">
         <CardHeader className="space-y-2">
           <div className="flex items-center justify-between gap-3">
@@ -81,24 +68,22 @@ export function AnnotationTaskDetailSkeleton({
         </CardFooter>
       </Card>
 
-      {shouldProcessFurther && (
-        <Card className="flex h-full flex-col overflow-hidden shadow-sm">
-          <CardHeader className="space-y-3">
-            <div className="space-y-1">
-              <Skeleton className="h-5 w-64" />
-              <Skeleton className="h-4 w-72" />
-            </div>
-          </CardHeader>
-          <CardContent className="flex-1 space-y-3 text-sm">
-            <Skeleton className="h-4 w-full" />
-            <Skeleton className="h-10 w-full rounded-md" />
-            <Skeleton className="h-4 w-full" />
-            <Skeleton className="h-10 w-full rounded-md" />
-            <Skeleton className="h-4 w-full" />
-            <Skeleton className="h-10 w-full rounded-md" />
-          </CardContent>
-        </Card>
-      )}
+      <Card className="flex h-full flex-col overflow-hidden shadow-sm">
+        <CardHeader className="space-y-3">
+          <div className="space-y-1">
+            <Skeleton className="h-5 w-64" />
+            <Skeleton className="h-4 w-72" />
+          </div>
+        </CardHeader>
+        <CardContent className="flex-1 space-y-3 text-sm">
+          <Skeleton className="h-4 w-full" />
+          <Skeleton className="h-10 w-full rounded-md" />
+          <Skeleton className="h-4 w-full" />
+          <Skeleton className="h-10 w-full rounded-md" />
+          <Skeleton className="h-4 w-full" />
+          <Skeleton className="h-10 w-full rounded-md" />
+        </CardContent>
+      </Card>
     </div>
   );
 }
